% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the CVS data file to read, see notes.}
}
\value{
A data frame with data readed from the CVS data file, or an error/stop if the
file does not exists.
}
\description{
The function reads data from a CVS file, stored on disk. The data are from
the \strong{US National Highway Traffic Safety Administration's}
\emph{Fatality Analysis Reporting System} (FARS), which is a nationwide
census providing the American public yearly data regarding fatal injuries
suffered in motor vehicle traffic crashes.
}
\details{
For more information, see:
\itemize{
  \item{\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}}
  \item{\url{https://en.wikipedia.org/wiki/Fatality_Analysis_Reporting_System}}.
}
}
\note{
To generate file name use: \code{\link{make_filename}}.
}
\examples{
\dontrun{
library(dplyr)
library(readr)
yr <- 2015
data <- yr \%>\%
  make_filename \%>\%
  fars_read
head(data)
}
}
\seealso{
\link{make_filename}
}
