% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS Data by Years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A string or integer vector with a list of years.}
}
\value{
A data.frame including entries in data by month and year, or NULL if the
\code{year} is not valid.
}
\description{
Auxiliary function used by \code{fars_summarize_years}. This function adds
the year as an additional column to the data.frame and select two columns
from that data.frame: month and year.
}
\examples{
fars_read_years(2013)
}
\seealso{
\link{fars_read}

\link{make_filename}

\link{fars_summarize_years}
}
