% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make FARS File Name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string or an integer with the input \code{year}.}
}
\value{
This function returns a string with the data file name for a given
  year, and the file path within the package.
}
\description{
Make a CVS data file name related to the given \code{year}.
The function does not check if the data file exists (is available).
}
\note{
To read file name use: \code{\link{fars_read}}
}
\examples{
make_filename(2013)
}
\seealso{
\link{fars_read}
}
